+++ include/gfastats-sak.h
@@ -193,7 +193,7 @@ public:
 
     bool compress(InSequences &inSequences, Instruction &instruction) {
         std::string *sequence = headerToSequence(instruction.contig1, inSequences, instruction);
-        std::vector<std::pair<unsigned int, unsigned int>> bedCoords;
+        std::vector<std::pair<unsigned long long int, unsigned long long int>> bedCoords;
         homopolymerCompress(sequence, bedCoords, instruction.compressThreshhold);
         compressStack.push(bedCoords);
 
@@ -206,7 +206,7 @@ public:
 
     bool decompress(InSequences &inSequences, Instruction &instruction) {
         std::string *sequence = headerToSequence(instruction.contig1, inSequences, instruction);
-        std::vector<std::pair<unsigned int, unsigned int>> bedCoords = compressStack.top();
+        std::vector<std::pair<unsigned long long int, unsigned long long int>> bedCoords = compressStack.top();
         compressStack.pop();
         homopolymerDecompress(sequence, bedCoords);
 
@@ -217,7 +217,7 @@ private:
     InSequences inSequences;
     InSegment inSegment1, inSegment2, inSegmentNew;
     std::string sId1Header, sId2Header;
-    std::stack<std::vector<std::pair<unsigned int, unsigned int>>> compressStack; // decompress uses indices of corresponding compresses, last compress = first decompress
+    std::stack<std::vector<std::pair<unsigned long long int, unsigned long long int>>> compressStack; // decompress uses indices of corresponding compresses, last compress = first decompress
 
     // unordered map from sak instruction to corresponding read and execute functions
     const phmap::flat_hash_map<std::string, Funcs> string_to_funcs {
